feat(editor): draw minimap marker lanes on whole pixels
The four right-anchored marker lanes were fractional shares of the 8 px
strip (8 / 6.56 / 5.12 / 3.68 px). They now round to the nearest whole
pixel, ties up, keeping the two-pixel floor: 8 / 7 / 5 / 4 px at
x = 0 / 1 / 3 / 4. Round-to-nearest is the integer scheme closest to the
old geometry (max 0.44 px, total 0.88 px deviation, against 0.68 / 1.36
for floor and 0.88 / 1.64 for ceil) and keeps four distinct nested lanes.

marker_lane_width and marker_lane_x are integer-only, the drawing code
converts to f64 once at the cairo call, and their two float_arithmetic
admissions are gone, so the minimap's recorded whole-pixel ceiling drops
from 10 to 8 in check-workflow-boundaries.py and the convention table.

OpenSpec change: integer-minimap-marker-lanes.

+++ crates/lushtext-core/src/ui/editor_page/minimap/projection_execution.rs
@@ -604,7 +604,7 @@ pub(super) fn draw_marker_strip(
     width: i32,
     height: i32,
 ) {
-    let width = f64::from(width.max(1));
+    let width = width.max(1);
     let dark = libadwaita::StyleManager::default().is_dark();
     let Some(source_map) = editor.imp().minimap.source_map.borrow().as_ref().cloned() else {
         return;
@@ -617,8 +617,14 @@ pub(super) fn draw_marker_strip(
         let (red, green, blue, alpha) = marker_rgba(marker.kind, dark);
         cr.set_source_rgba(red, green, blue, alpha);
         // Cairo takes a height; the policy hands over edges, so the one
-        // subtraction of two fractional widget coordinates happens here.
-        cr.rectangle(x, marker.top, lane_width, marker.bottom - marker.top);
+        // subtraction of two fractional widget coordinates happens here. The
+        // lane is whole-pixel and becomes `f64` only at this cairo call.
+        cr.rectangle(
+            f64::from(x),
+            marker.top,
+            f64::from(lane_width),
+            marker.bottom - marker.top,
+        );
         let _ = cr.fill();
     }
 }
